pyproject.toml
[tool.poetry]
name = "drf-proj"
version = "0.1.0"
description = ""
authors = ["georgii <goqwertys@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.10"
pillow = "^11.0.0"
django-phonenumber-field = "^8.0.0"
phonenumbers = "^8.13.50"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
